playing with different Widgets
I'm trying to find a way to implement what we would like to have. Most
likely a kind of StackedLayout or StackedWidget.

I don't know yet if those already-available widgets will be easier to
use than running our own kind of super-container. Most likely running
our own will be easier in the end (even if the end is close).